src/datawizard/interfaces/data_interface.c: fix test

+++ src/datawizard/interfaces/data_interface.c
@@ -439,7 +439,7 @@ void starpu_data_register_ops(struct starpu_data_interface_ops *ops)
 
 void starpu_data_register(starpu_data_handle_t *handleptr, int home_node, void *data_interface, struct starpu_data_interface_ops *ops)
 {
-	STARPU_ASSERT_MSG(home_node >= -1 && home_node < starpu_memory_nodes_get_count(), "Invalid memory node number");
+	STARPU_ASSERT_MSG(home_node >= -1 && home_node < (int)starpu_memory_nodes_get_count(), "Invalid memory node number");
 	starpu_data_handle_t handle = _starpu_data_handle_allocate(ops, home_node);
 
 	STARPU_ASSERT(handleptr);
